requirements.txt
# Core Dependencies
web3>=6.0.0
python-dotenv>=1.0.0

# Database & Persistence (sqlite3 is built-in)
sqlalchemy>=2.0.0
alembic>=1.12.0

# Retry Logic & Error Handling
tenacity>=8.2.0
backoff>=2.2.0

# Structured Logging
structlog>=23.1.0
colorama>=0.4.6

# Metrics & Monitoring
prometheus-client>=0.17.0

# Data Validation (Pydantic V2)
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Multi-Channel Notifications
apprise>=1.4.0

# Async Operations
aiofiles>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Optional: Redis for caching
redis>=4.6.0

# Optional: Dashboard
streamlit>=1.25.0
plotly>=5.15.0 